+++ src/burn/engine/msiengine.h
@@ -51,6 +51,11 @@ HRESULT MsiEnginePlanAddPackage(
     __in_opt HANDLE hCacheEvent,
     __in BOOL fPlanPackageCacheRollback
     );
+HRESULT MsiEngineAddCompatiblePackage(
+    __in BURN_PACKAGES* pPackages,
+    __in const BURN_PACKAGE* pPackage,
+    __out_opt BURN_PACKAGE** ppCompatiblePackage
+    );
 HRESULT MsiEngineExecutePackage(
     __in_opt HWND hwndParent,
     __in BURN_EXECUTE_ACTION* pExecuteAction,